+++ src/main/java/com/example/rentsystem/UsersRepository.java
@@ -1,12 +1,10 @@
 package com.example.rentsystem;
 
+import com.example.user.UserType;
 import org.springframework.data.repository.CrudRepository;
 
 public interface UsersRepository extends CrudRepository<User, Integer> {
 //    User findByUsername(String username);
-}
-
-interface UserTypeRepository extends CrudRepository<UserType, Integer> {
 }
 
 interface VehicleOwnerRepository extends CrudRepository<VehicleOwner, Integer> {
