+++ src/main/java/com/example/user/AppUser.java
@@ -9,8 +9,6 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-import com.example.rentsystem.UserType;
-
 import javax.persistence.*;
 
 @Entity
