+++ src/main/java/com/example/user/AppUserRepository.java
@@ -2,10 +2,12 @@ package com.example.user;
 
 //import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 
 public interface AppUserRepository extends JpaRepository<AppUser, Integer> {
     AppUser findByName(String name);
 }
 
-//interface UserTypeRepository extends CrudRepository<UserType, Integer> {
-//}
+interface UserTypeRepository extends CrudRepository<UserType, Integer> {
+}
+
