+++ src/main/java/com/example/user/UserController.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 @RestController
 @RequestMapping("/users")
 public class UserController {
-//    @Autowired
+    @Autowired
     private AppUserRepository userRepository;
 
     private BCryptPasswordEncoder bCryptPasswordEncoder;
@@ -21,14 +21,6 @@ public class UserController {
         this.bCryptPasswordEncoder = bCryptPasswordEncoder;
     }
 
-//    @PostMapping("/sign-up")
-//    public AppUser signUp(@RequestBody AppUser appUser) {
-//        System.out.println("sign-up called...");
-////        appUser.setPassword(bCryptPasswordEncoder.encode(appUser.getPassword()));
-//        appUser.setPassword(appUser.getPassword());
-//        userRepository.save(appUser);
-//        return appUser;
-//    }
     @PostMapping("/sign-up")
     public AppUser signUp(@RequestBody AppUser appUser) {
         System.out.println("sign-up called...");
